+++ lyscripts/compute/priors.py
@@ -17,9 +17,10 @@ from lyscripts.configs import (
     DistributionConfig,
     GraphConfig,
     ModelConfig,
-    add_dists,
+    add_distributions,
     construct_model,
 )
+from lyscripts.utils import console
 
 
 def compute_priors(
@@ -39,13 +40,14 @@ def compute_priors(
     each of the ``t_stages`` and marginalized over using the ``t_stages_dist``.
     """
     model = construct_model(model_config, graph_config)
-    model = add_dists(model, dist_configs)
+    model = add_distributions(model, dist_configs)
     priors = []
 
     for sample in progress.track(
         sequence=samples,
-        description="[blue]INFO     [/blue]" + progress_desc,
+        description=progress_desc,
         total=len(samples),
+        console=console,
     ):
         model.set_params(*sample)
         priors.append(