+++ lyscripts/compute/risks.py
@@ -20,10 +20,11 @@ from lyscripts.configs import (
     InvolvementConfig,
     ModalityConfig,
     ModelConfig,
-    add_dists,
+    add_distributions,
     add_modalities,
     construct_model,
 )
+from lyscripts.utils import console
 
 
 def compute_risks(
@@ -42,14 +43,15 @@ def compute_risks(
     distribution over the states that correspond to the involvement of interest.
     """
     model = construct_model(model_config, graph_config)
-    model = add_dists(model, dist_configs)
+    model = add_distributions(model, dist_configs)
     model = add_modalities(model, modality_configs)
     risks = []
 
     for posterior in progress.track(
         sequence=posteriors,
-        description="[blue]INFO     [/blue]" + progress_desc,
+        description=progress_desc,
         total=len(posteriors),
+        console=console,
     ):
         risks.append(
             model.marginalize(involvement=involvement, given_state_dist=posterior)